DjangoDashBoard/Telegram_bot/requirements.txt
asgiref==3.8.1
Django==5.2.3
dotenv==0.9.9
mysqlclient==2.2.7
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
anyio==4.9.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
mysql-connector-python==9.3.0
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1